Refatora casos de uso Codaf e adiciona testes unitários
AB#140016

+++ SME.ConectaFormacao.Infra.Dados/Repositorios/Interfaces/IRepositorioCodafListaPresenca.cs
@@ -5,6 +5,6 @@ namespace SME.ConectaFormacao.Infra.Dados.Repositorios.Interfaces
 {
     public interface IRepositorioCodafListaPresenca : IRepositorioBaseAuditavel<CodafListaPresenca>
     {
-        Task<bool> TurmaJaTemListaDePresencaAsync(long propostaTurmaId, int listaPresencaId = 0);
+        Task<bool> TurmaJaTemListaDePresencaAsync(long propostaTurmaId, long listaPresencaId = 0);
     }
 }
